refactor(scripts): move legacy scanner extraction modules
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
